Pie Chart Integration
Added budgeting Pie Chart to display spending. The legend includes USD spent on each slice. Count reports of each purchase are displayed on the terminal. Removed the line graph feature for the account balance. A standalone executable for this program will be available once the manual entry sequence can be executed more smoothly, likely with a GUI. 18 functions show each category that will be tracked. Each contains a list of names that the program will use to sort purchases depending on the description. This is used to create the Pie Chart.